+++ sdk/src/main/java/com/mapbox/maps/renderer/MapboxWidgetRenderer.kt
@@ -14,7 +14,7 @@ internal class MapboxWidgetRenderer(
   private val antialiasingSampleCount: Int,
 ) {
   private var eglCore: EGLCore? = null
-  private var eglPrepared = false
+  private var eglContextCreated = false
   private var eglSurface: EGLSurface? = null
   private var sizeChanged = false
 
@@ -36,7 +36,7 @@ internal class MapboxWidgetRenderer(
   fun getTexture() = textures[0]
 
   fun setSharedContext(sharedContext: EGLContext) {
-    if (eglPrepared) {
+    if (eglContextCreated) {
       release()
     }
     eglCore = EGLCore(
@@ -120,7 +120,7 @@ internal class MapboxWidgetRenderer(
     }
     this.eglSurface = null
     this.eglCore = null
-    this.eglPrepared = false
+    this.eglContextCreated = false
   }
 
   fun updateTexture() {
@@ -177,9 +177,9 @@ internal class MapboxWidgetRenderer(
       return
     }
 
-    if (!eglPrepared) {
-      eglPrepared = eglCore.prepareEgl()
-      if (!eglPrepared) {
+    if (!eglContextCreated) {
+      eglContextCreated = eglCore.prepareEgl()
+      if (!eglContextCreated) {
         logE(TAG, "Widget EGL was not configured, please check logs above.")
         return
       }